added server name via two-way databinding
Difference? in the previous commit, event binding. I was able to be
specific on which property I wanted to bind (input). and then pass the
data that input sends us to the variable declared in the typsecript
file and thus update the html <p> via string interpolation.

two-way binding constrains you to binding to ngModel, but not only will
it bind to your typescript variable. It will now also bind to the input
property in your InputElement.